Express error handling

+++ processors/src/claim-processor/handler.js
@@ -78,6 +78,11 @@ class ClaimHandler extends TransactionHandler {
    */
   async deleteClaim(claim, state) {
     const { vin } = claim.vehicle;
+    const existingClaim = await state.getClaim(vin);
+    if (!existingClaim) {
+      logger.error(`Attempted to delete a non-existing claim for VIN: ${vin}.`);
+      throw new InvalidTransaction('Cannot delete - claim does not exist');
+    }
     return state.deleteClaim(vin);
   }
 
@@ -92,7 +97,7 @@ class ClaimHandler extends TransactionHandler {
     const existingClaim = await state.getClaim(vin);
     if (!existingClaim) {
       logger.error(`Attempted to edit a non-existing claim for VIN: ${vin}.`);
-      throw new InvalidTransaction('Claim does not exist');
+      throw new InvalidTransaction('Cannot edit - claim does not exist');
     }
     logger.debug(`Editing existing claim using VIN: ${vin}`);
     return state.setClaim(vin, {
